Add home page in new client
