+++ app/views/Home/RecentActivity/ActivityItem/styles.css
