+++ app/views/Home/RecentActivity/styles.css
@@ -14,7 +14,6 @@
         flex-direction: column;
         flex-grow: 1;
         justify-content: center;
-        // color: var(--dui-color-text-watermark);
         color: var(--dui-color-text-disabled);
 
         .icon {
