--- src/newViews/Home/Summary/styles.scss
@@ -1,31 +1,0 @@
-.summary {
-    --gap: var(--dui-spacing-large);
-    --max-item-width: 20rem;
-    --min-item-width: 13rem;
-    --item-per-row: 2;
-    --ipr-1: calc(var(--item-per-row) - 1);
-
-    position: relative;
-    min-width: calc(var(--item-per-row) * var(--min-item-width) + var(--gap) * var(--ipr-1) + var(--dui-spacing-large) * 2);
-    max-width: calc(var(--item-per-row) * var(--max-item-width) + var(--gap) * var(--ipr-1) + var(--dui-spacing-large) * 2);
-    overflow: auto;
-
-    .content {
-        display: flex;
-        flex-wrap: wrap;
-        justify-content: space-between;
-        overflow: auto;
-        gap: var(--gap);
-
-        >* {
-            flex-basis: calc((100% - (var(--gap) * var(--ipr-1))) / var(--item-per-row));
-            min-width: var(--min-item-width);
-            max-width: var(--max-item-width);
-        }
-    }
-
-    .percentage-info-card {
-        padding-right: calc(var(--dui-spacing-medium) - var(--dui-spacing-small));
-        padding-left: calc(var(--dui-spacing-medium) - var(--dui-spacing-small));
-    }
-}
