--- src/newViews/Home/styles.scss
@@ -1,107 +1,0 @@
-.home {
-    display: flex;
-    background-color: var(--dui-color-background);
-    height: 100%;
-    overflow-x: hidden;
-    overflow-y: auto;
-
-    .main-content {
-        flex-grow: 1;
-        padding-left: var(--dui-spacing-large);
-
-        .top-section {
-            display: flex;
-            flex-wrap: wrap;
-            padding: 0 calc(var(--dui-spacing-large) - var(--dui-spacing-medium));
-
-            .summary-container {
-                flex-basis: 10%;
-                flex-grow: 1;
-                overflow: auto;
-            }
-
-            .tagging-activity-container {
-                display: flex;
-                flex-basis: 40%;
-                flex-direction: column;
-                flex-grow: 1;
-                min-width: 30rem;
-                overflow: auto;
-
-                .content {
-                    display: flex;
-                    flex-direction: column;
-                    flex-grow: 1;
-                    overflow: auto;
-
-                    .tagging-activity {
-                        flex-grow: 1;
-                        overflow: auto;
-                    }
-                }
-            }
-        }
-
-        .bottom-section {
-            padding: 0 calc(var(--dui-spacing-large) - var(--dui-spacing-medium));
-
-            .bottom-section-content {
-                position: relative;
-                padding: 0 var(--dui-spacing-medium);
-
-                .project-item {
-                    margin: var(--dui-spacing-medium) 0;
-                }
-            }
-        }
-
-        .empty-recent-project {
-            display: flex;
-            align-items: center;
-            flex-direction: column;
-            flex-grow: 1;
-            justify-content: center;
-            margin: var(--dui-spacing-large) 0;
-            background-color: var(--dui-color-foreground);
-            padding: var(--dui-spacing-extra-large) var(--dui-spacing-large);
-            min-height: 30vh;
-            // color: var(--dui-color-text-watermark);
-            color: var(--dui-color-text-disabled);
-
-            .icon {
-                flex-shrink: 0;
-                padding: var(--dui-spacing-small);
-                font-size: 10em;
-            }
-
-            .text {
-                padding: var(--dui-spacing-small);
-                text-align: center;
-                font-size: var(--dui-font-size-large);
-            }
-        }
-    }
-
-    .side-content {
-        align-self: flex-start;
-        flex-basis: calc(12em + 10vw);
-        flex-grow: 0;
-        flex-shrink: 0;
-        background-color: var(--dui-color-foreground);
-        padding-right: var(--dui-spacing-large);
-        overflow: auto;
-    }
-}
-
-.select-option {
-    display: flex;
-    justify-content: space-between;
-
-    .badge {
-        padding-right: 0;
-
-        &.active {
-            color: var(--color-text-on-accent);
-        }
-    }
-}
